Refactors the talkie controller and move it to a concern
