+++ lib/talkie.rb
@@ -18,3 +18,4 @@ require "talkie/acts_as_talker"
 require "talkie/acts_as_commentable"
 require "talkie/comments_renderer_helper"
 require "talkie/permission"
+require "talkie/controller"
