+++ lib/talkie/engine.rb
@@ -7,6 +7,7 @@ module Talkie
     initializer 'talkie_helpers.action_controller' do
       ActiveSupport.on_load :action_controller do
         helper Talkie::CommentsHelper
+        include Talkie::Controller
       end
     end
   end
