+++ spec/dummy/app/assets/javascripts/application.js
@@ -1,0 +1,1 @@
+//= require rails-ujs
